refactor(perl-test-generators): split code generators into srp modules (#0000)
